requirements.txt
Django==5.0.7
djangorestframework
drf-yasg
requests

psycopg2-binary
